backend/requirements.txt
fastapi
uvicorn
httpx
groq
python-dotenv
rapidfuzz
python-dateutil
pydantic